improved image upload prompt and added auth to more routes

+++ main.py
@@ -1,7 +1,8 @@
 import uvicorn
 
-from fastapi import FastAPI, UploadFile, status, HTTPException
+from fastapi import Depends, FastAPI, Request, UploadFile, status, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
+from sqlalchemy.orm import Session
 import os
 import google.generativeai as genai
 import PIL
@@ -13,6 +14,7 @@ load_dotenv()
 from app.routes.users import user_router
 from app.routes.auth import router
 from app.db import Base, engine, SessionLocal
+from app.auth.auth_handler import get_current_user
 
 Base.metadata.create_all(bind=engine)
 
@@ -22,6 +24,7 @@ def get_db():
         yield db
     except:
         db.close()
+        raise
 
 
 app = FastAPI()
@@ -42,13 +45,21 @@ genai.configure(api_key=os.getenv("GOOGLE_GEMINI_API_KEY"))
 model = genai.GenerativeModel("gemini-1.5-flash")
 
 @app.post("/upload")
-async def sketch_to_code(file: UploadFile):
+async def sketch_to_code(file: UploadFile, req: Request, db: Session= Depends(get_db)):
+    if not req.headers["authorization"].split(" ")[1]:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate",
+            headers={"WWW-Authenticate": "Bearer"}
+        )
+    token = req.headers["authorization"].split(" ")[1]
+    await get_current_user(token, db)
     try:
         os.mkdir("uploads")
         image_bytes = await file.read()
         image = PIL.Image.open(io.BytesIO(image_bytes))
         image.save("uploads/" + file.filename)
-        response = model.generate_content(["The image data that is provided to you will contain an image or rough sketch of a website, so you need to provide html, css and js code for that website and also keep in mind that css and js code should be inside that html tag only. Incase an image data can't be converted to html code then respond as follows:- 'Provide a proper image which contains information about website'", image])
+        response = model.generate_content(["Convert the image into HTML, CSS and JavaScript equivalent code only the frontend or UI not the backend. If the image can't be converted, return 'Failed to process image'. Use internal CSS and Javascript, add proper styling, include all the features, Make the code fully functional.", image])
         print(response.text)
         os.remove(f"uploads/{file.filename}")
         os.rmdir("uploads")
